refactor code and avoid circular imports
