+++ src/xlnstorch/__init__.py
@@ -18,6 +18,11 @@ from .dispatch_table import (
     override_implementation,
     apply_lns_op
 )
+from .autograd import LNSFunction
+from .tensor_utils import (
+    align_lnstensor_bases,
+    format_lnstensor_operands,
+)
 from .tensor import (
     LNSTensor,
     lnstensor,
@@ -31,9 +36,6 @@ from .tensor import (
     rand_like,
     randn,
     randn_like,
-    LNSFunction,
-    align_lnstensor_bases,
-    format_lnstensor_operands,
 )
 from . import operators
 from . import nn
