+++ src/xlnstorch/operators/activation_ops.py
@@ -1,5 +1,6 @@
 import torch
-from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, rand, LNSFunction
+from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, rand
+from ..autograd import LNSFunction
 from . import (
     lns_mul,
     lns_add,
