+++ src/xlnstorch/operators/arithmetic_ops.py
@@ -1,5 +1,6 @@
 import torch
-from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, full_like, LNSFunction
+from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, full_like
+from ..autograd import LNSFunction
 from . import (
     lns_add,
     lns_neg,
