+++ src/xlnstorch/operators/layer_ops.py
@@ -1,7 +1,8 @@
 import warnings
 
 import torch
-from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, zeros, zeros_like, LNSFunction
+from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, zeros, zeros_like
+from ..autograd import LNSFunction
 from . import (
     lns_mul,
     lns_sum,
